encryptEncryptionKey method is exposed to the user.
